refactor(hydro_deploy): moved launch compute engine into shared code between gcp and azure
