+++ hydro_deploy/core/src/lib.rs
@@ -10,6 +10,8 @@ use hydroflow_cli_integration::ServerBindConfig;
 use tokio::sync::RwLock;
 
 pub mod deployment;
+pub mod launch;
+pub use launch::LaunchedVirtualMachine;
 pub use deployment::Deployment;
 
 pub mod progress;
